+++ entity/src/controllers/mod.rs
@@ -1,2 +1,2 @@
-pub mod chatbot;
-pub mod chatbot_setting;
+pub mod chatbot_controller;
+pub mod chatbot_setting_controller;
