+++ entity/src/models/chatbot_model.rs
