+++ entity/src/models/chatbot_setting_model.rs
