+++ entity/src/models/mod.rs
@@ -1,5 +1,5 @@
-pub mod chatbot_setting;
-pub mod chatbot;
+pub mod chatbot_setting_model;
+pub mod chatbot_model;
 pub mod failed_job;
 pub mod jobs;
 pub mod pdf_data_source;
