+++ entity/src/schemas/chatbot_schema.rs
