+++ entity/src/schemas/chatbot_settings_schema.rs
@@ -2,7 +2,7 @@ use chrono::NaiveDateTime;
 use serde::{Serialize, Deserialize};
 
 #[derive(Debug, Serialize, Deserialize)]
-pub struct ChatbotSettingSchema {
+pub struct CreateChatbotSetting {
     pub id: String,
     pub chatbot_id: String,
     pub name: String,
